session_test.rs: Move some tests to integration
