+++ scylla/src/lib.rs
@@ -82,7 +82,7 @@
 //!     .query_unpaged("SELECT a, b FROM ks.tab", &[])
 //!     .await?
 //!     .into_rows_result()?;
-//!     
+//!
 //! for row in query_rows.rows()? {
 //!     // Parse row as int and text \
 //!     let (int_val, text_val): (i32, &str) = row?;
@@ -222,16 +222,6 @@ pub mod deserialize {
         };
     }
 
-    // Shorthands for better readability.
-    #[cfg_attr(not(test), allow(unused))]
-    pub(crate) trait DeserializeOwnedValue:
-        for<'frame, 'metadata> DeserializeValue<'frame, 'metadata>
-    {
-    }
-    impl<T> DeserializeOwnedValue for T where
-        T: for<'frame, 'metadata> DeserializeValue<'frame, 'metadata>
-    {
-    }
     pub(crate) trait DeserializeOwnedRow:
         for<'frame, 'metadata> DeserializeRow<'frame, 'metadata>
     {
