+++ scylla/tests/integration/main.rs
@@ -13,6 +13,7 @@ mod lwt_optimisation;
 mod new_session;
 mod retries;
 mod self_identity;
+mod session;
 mod shards;
 mod silent_prepare_batch;
 mod silent_prepare_query;
